TinyEngine: Move private header files into "src" directory
Only keep headers in the "include" directory which are meant to be included by users
of this library. Move the rest of the header files into the "src" directory.

While at it change multi-line macros to be inside a "do { } while (0)" do allow
";" at the line end when the macro is used. And correct the include guard of
arm_nnfunctions_modified.h.

Signed-off-by: Henrik Claßen <h.classen@campus.tu-berlin.de>